Created auth\Login and moved Login Logic in there
